Needed to extract out the interface, or else the CacheFactory did not wire up.
